+++ src/array/codec/array_to_bytes.rs
@@ -3,7 +3,7 @@
 pub mod bytes;
 pub mod codec_chain;
 pub mod vlen;
-pub mod vlen_interleaved;
+pub mod vlen_v2;
 
 #[cfg(feature = "pcodec")]
 pub mod pcodec;