+++ src/metadata/v3.rs
@@ -29,8 +29,8 @@ pub mod codec {
     pub mod transpose;
     /// `vlen` codec metadata.
     pub mod vlen;
-    /// `vlen_interleaved` codec metadata.
-    pub mod vlen_interleaved;
+    /// `vlen_v2` codec metadata.
+    pub mod vlen_v2;
     #[cfg(feature = "zfp")]
     /// `zfp` codec metadata.
     pub mod zfp;